requirements.txt
streamlit
langchain
langchain_openai
langchain_community
langchain-chroma
chromadb
python-dotenv
unstructured
pdf2image
pytesseract
langchain-core
python-docx
pdfminer.six
python-pptx
beautifulsoup4
pandas
plotly
pydantic
pydantic-settings
uuid
scikit-learn
fastapi
uvicorn
pillow
openai
openpyxl
sentence-transformers
rank-bm25
nltk
transformers
torch
